+++ QuoteDroplet/Views/SubViews/AuthorView.swift
