+++ QuoteDroplet/Views/SubViews/SingleQuoteView.swift
